# Conflicts:
#	tests/test_other.py

+++ tests/core/test_i32_mul_precise.c
@@ -6,14 +6,15 @@
  */
 
 #include <stdio.h>
+#include <stdint.h>
 
 int main(int argc, char **argv) {
-  unsigned long d1 = 0x847c9b5d;
-  unsigned long q = 0x549530e1;
+  uint32_t d1 = 0x847c9b5d;
+  uint32_t q = 0x549530e1;
   if (argc > 1000) {
     q += argc;
     d1 -= argc;
   }  // confuse optimizer
-  printf("%lu\n", d1 * q);
+  printf("%u\n", d1 * q);
   return 0;
 }
